GroceryMart_api/requirements.txt
asgiref==3.9.0
attrs==25.3.0
black==25.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.8.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
responses==0.25.7
rpds-py==0.27.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
sslcommerz-lib==1.0
sslcommerz_python_api==1.0.1
stripe==12.4.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
